move tests to `test/` dir
